+++ app/contact/contact.hpp
@@ -10,9 +10,9 @@ class Contact : public CppAppContainer
     
     static void loadContacts(bool force = false);
     static void saveContacts();
-    static uint contactPage();
-    static void editContact(bool create = false, uint index = -1);
-    static void showContact(uint index);
+    static uint16_t contactPage();
+    static void editContact(bool create = false, uint16_t index = -1);
+    static void showContact(uint16_t index);
 
     struct OneContact
     {
